+++ src/main/java/br/com/springweb/entities/Cidade.java
@@ -16,7 +16,7 @@ public class Cidade {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private String nome;
 
 	@JsonManagedReference
@@ -28,18 +28,18 @@ public class Cidade {
 
 	}
 
-	public Cidade(Long id, String nome, Estado estado) {
+	public Cidade(Integer id, String nome, Estado estado) {
 		super();
 		this.id = id;
 		this.nome = nome;
 		this.estado = estado;
 	}
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
