+++ src/main/java/br/com/springweb/repositorys/ProdutoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 import br.com.springweb.entities.Produto;
 
 @Repository
-public interface ProdutoRepository extends JpaRepository<Produto,Long> {
+public interface ProdutoRepository extends JpaRepository<Produto,Integer> {
 
 	
 }
